requirements.txt
cssselect==0.9.1
lxml==3.6.0
pyquery==1.2.13
requests==2.10.0
unicode-slugify==0.1.3
